--- packages/hydrated_bloc/test/mocks/mocks.dart
@@ -1,1 +1,0 @@
-export 'mock_path_provider_platform.dart';
